Patch 0.1: - added dashboard page - move old:dashboard class and file to transaction - added export to csv
