+++ app/Models/Transaction.php
@@ -5,7 +5,7 @@ namespace App\Models;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
-class Dashboard extends Model
+class Transaction extends Model
 {
     use HasFactory;
     public $table = "pembelian";
