+++ composer.json
@@ -10,6 +10,8 @@
         "laravel/framework": "^10.10",
         "laravel/sanctum": "^3.2",
         "laravel/tinker": "^2.8",
+        "phpfastcache/phpfastcache": "^9.1",
+        "raiym/instagram-php-scraper": "^0.5.0",
         "yajra/laravel-datatables": "^10.0",
         "yajra/laravel-datatables-oracle": "^10.3.1"
     },
