+++ resources/views/etc/navbar.blade.php
@@ -32,15 +32,7 @@
                         @auth
                             @if (auth()->user()->role == 'admin')
                                 <li>
-                                    <a href="manage"
-                                        class="flex items-center p-2 text-gray-900 rounded-lg hover:bg-gray-100 group">
-                                        <i
-                                            class="flex-shrink-0 w-5 h-5 text-gray-500 transition duration-75 dark:text-gray-400 group-hover:text-gray-900 fa-solid fa-house"></i>
-                                        <span class="ml-3">Home</span>
-                                    </a>
-                                </li>
-                                <li>
-                                    <a href="/manage/dashboard"
+                                    <a href="/manage"
                                         class="flex items-center p-2 text-gray-900 rounded-lg hover:bg-gray-100 group">
                                         <svg class="w-5 h-5 text-gray-500 transition duration-75 dark:text-gray-400 group-hover:text-gray-900"
                                             aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="currentColor"
@@ -51,6 +43,13 @@
                                                 d="M12.5 0c-.157 0-.311.01-.565.027A1 1 0 0 0 11 1.02V10h8.975a1 1 0 0 0 1-.935c.013-.188.028-.374.028-.565A8.51 8.51 0 0 0 12.5 0Z" />
                                         </svg>
                                         <span class="ml-3">Dashboard</span>
+                                    </a>
+                                </li>
+                                <li>
+                                    <a href="/manage/tx"
+                                       class="flex items-center p-2 text-gray-900 rounded-lg hover:bg-gray-100 group">
+                                        <i class="flex-shrink-0 w-5 h-5 text-gray-500 transition duration-75 dark:text-gray-400 group-hover:text-gray-900 fa-solid fa-receipt"></i>
+                                        <span class="flex-1 ml-3 whitespace-nowrap">Transaction</span>
                                     </a>
                                 </li>
                                 <li>
